fix: added skillProvider
